Roleplay_doh/requirements.txt
datasets>=2.14.0
python-dotenv>=1.0.0
pandas>=1.5.0
rapidfuzz>=3.0.0
python-json-logger>=2.0.0
